feat: some improvements in security

+++ src/main/java/br/com/les/amstore/controller/AddressController.java
@@ -4,11 +4,15 @@ import br.com.les.amstore.domain.Address;
 import br.com.les.amstore.domain.Customer;
 import br.com.les.amstore.domain.State;
 import br.com.les.amstore.dto.AddressDTO;
-import br.com.les.amstore.service.*;
+import br.com.les.amstore.service.IAddressService;
+import br.com.les.amstore.service.IAddressTypeService;
+import br.com.les.amstore.service.ICustomersService;
+import br.com.les.amstore.service.IStateService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
@@ -18,7 +22,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Controller
-//@RequestMapping("/address")
 @RequestMapping("/customer/edit")
 public class AddressController {
 
@@ -54,6 +57,7 @@ public class AddressController {
         mv.addObject(customer);
         mv.addObject(address);
 
+        customers.isCurrentUserLoggedIn(customer.getId(), mv);
         return mv;
     }
 
@@ -66,12 +70,16 @@ public class AddressController {
         mv.addObject(customer);
         mv.addObject(address);
 
+        customers.isCurrentUserLoggedIn(customer.getId(), mv);
         return mv;
     }
 
     @PostMapping("/{id}/addresses/new")
     public ModelAndView createCustomerAddress(@PathVariable("id") Customer customer, @Valid Address address, BindingResult result, RedirectAttributes attributes) {
         ModelAndView mv = new ModelAndView("redirect:/customer/edit/" + customer.getId() + "/addresses/new");
+
+        if(!customers.isCurrentUserLoggedIn(customer.getId()))
+            result.addError(new ObjectError("customer", "Você não está autorizado a executar essa ação"));
 
         if(result.hasErrors() || null == customer.getId()){
             System.err.println("Deu erro");
@@ -97,15 +105,17 @@ public class AddressController {
         mv.addObject(customer);
         mv.addObject(address);
 
+        customers.isCurrentUserLoggedIn(customer.getId(), mv);
         return mv;
     }
 
     @PostMapping(value = "/{customer_id}/addresses/{id}/edit")
     public ModelAndView updateAddress(@PathVariable(value = "customer_id") Customer customer, @Valid Address address, BindingResult result, RedirectAttributes attributes) {
-        if(result.hasErrors() || null == customer.getId()){
-            System.err.println("Deu erro");
+        if(!customers.isCurrentUserLoggedIn(customer.getId()))
+            result.addError(new ObjectError("customer", "Você não está autorizado a executar essa ação"));
+
+        if(result.hasErrors() || null == customer.getId())
             return newAddress(customer, address);
-        }
 
         customer = customers.findById(customer.getId());
 
@@ -120,7 +130,10 @@ public class AddressController {
     }
 
     @DeleteMapping("/{id}/addresses")
-    public ModelAndView deleteAddress(@RequestParam String id, @RequestParam String address_id, RedirectAttributes attributes) {
+    public ModelAndView deleteAddress(@PathVariable(value = "id") Customer customer, @RequestParam String id, @RequestParam String address_id, RedirectAttributes attributes) {
+        if(!customers.isCurrentUserLoggedIn(customer.getId()))
+            return new ModelAndView("redirect:/");
+
         Address address = addresses.findById(Long.parseLong(address_id));
         address.delete();
